Litt bedre navngivning
